app/requirements.txt
python-dotenv>=0.19.0
aiofiles>=0.8.0
python-telegram-bot==21.4
deepgram-sdk>=3.0.0
httpx==0.27.2